+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,7 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+- limit citatory query to 10 items due to ops restriction
 
 0.11.0
 ======