+++ elmyra.ip.access.epo/CHANGES.rst
@@ -7,6 +7,7 @@ development
 - limit citatory query to 10 items due to ops restriction
 - ui: modify/extend citation link buttons
 - ui: "new project" action
+- ui: display number of entries in basket
 
 0.11.0
 ======