+++ elmyra.ip.access.epo/CHANGES.rst
@@ -8,6 +8,7 @@ development
 - ui: modify/extend citation link buttons
 - ui: "new project" action
 - ui: display number of entries in basket
+- ui: improve classification display: add ipc and cpc classes
 
 0.11.0
 ======